+++ concordium-sdk/src/main/java/com/concordium/sdk/ClientV2.java
@@ -281,7 +281,7 @@ public final class ClientV2 {
 
 
     /**
-     * Get the list of transactions hashes for transactions that claim to be from the given account,
+     * Get the list of transactions hashes for transactions from the given account,
      * but which are not yet finalized. They are either committed to a block or still pending.
      * If the account does not exist an empty list will be returned.
      *